+++ samples/Profile/ReferencesCodeSample.php
@@ -18,16 +18,69 @@ $auth = new \idOS\Auth\CredentialToken(
 $sdk = \idOS\SDK::create($auth);
 
 /**
- * Calling the Profile Endpoint passing the username, and after that, the References Endpoint and the method listAll.
+ * Creates a new reference.
  */
 $response = $sdk
     ->Profile($credentials['username'])
-    ->References->listAll();
+    ->References->createNew('reference', 'value');
 
-/**
- * Prints the response.
- */
-print_r($response);
+if ($response['status'] === true) {
+
+    /**
+     * Calling the Profile Endpoint passing the username, and after that, the References Endpoint and the method listAll.
+     */
+    $response = $sdk
+        ->Profile($credentials['username'])
+        ->References->listAll();
+
+    /**
+     * Prints the response.
+     */
+    foreach ($response['data'] as $reference) {
+        print_r("\nName: " . $reference['name']);
+        print_r("\nValue: " . $reference['value']);
+        print_r("\n");
+    }
+
+    /**
+     * Updates a reference.
+     */
+    $response = $sdk
+        ->Profile($credentials['username'])
+        ->References->updateOne('reference', 'new-value');
+
+    /**
+     * Prints the response.
+     */
+    print_r("\nName: " . $response['data']['name']);
+    print_r("\nValue: " . $response['data']['value']);
+    print_r("\n");
+
+    /**
+     * Retrieves the reference updated.
+     */
+    $response = $sdk
+        ->Profile($credentials['username'])
+        ->References->getOne($response['data']['name']);
+
+    /**
+     * Prints the response.
+     */
+    print_r("\nName: " . $response['data']['name']);
+    print_r("\nValue: " . $response['data']['value']);
+    print_r("\n");
+    /**
+     * Deletes the reference created/updated.
+     */
+    $response = $sdk
+        ->Profile($credentials['username'])
+        ->References->deleteOne($response['data']['name']);
+    /**
+     * Prints the response status.
+     */
+    print_r("\nStatus: " . $response['status']);
+    print_r("\n");
+}
 
 /**
  * Creates a new reference.
@@ -36,54 +89,17 @@ $response = $sdk
     ->Profile($credentials['username'])
     ->References->createNew('reference', 'value');
 
-/**
- * Prints the response.
- */
-print_r($response);
+if ($response['status'] === true) {
+    /**
+     * Deletes all references.
+     */
+    $response = $sdk
+        ->Profile($credentials['username'])
+        ->References->deleteAll();
 
-/**
- * Updates a reference.
- */
-$response = $sdk
-    ->Profile($credentials['username'])
-    ->References->updateOne('reference', 'new-value');
-
-/**
- * Prints the response.
- */
-print_r($response);
-
-/**
- * Retrieves the reference updated.
- */
-$response = $sdk
-    ->Profile($credentials['username'])
-    ->References->getOne('reference');
-
-/**
- * Prints the response.
- */
-print_r($response);
-
-/**
- * Deletes the reference created/updated.
- */
-$response = $sdk
-    ->Profile($credentials['username'])
-    ->References->deleteOne('reference');
-/**
- * Prints the response.
- */
-print_r($response);
-
-/**
- * Deletes all references.
- */
-$response = $sdk
-    ->Profile($credentials['username'])
-    ->References->deleteAll();
-
-/**
- * Prints the response.
- */
-print_r($response);
+    /**
+     * Prints the number of deleted references.
+     */
+    print_r("\nDeleted references: " . $response['deleted']);
+    print_r("\n");
+}
